Automatic Reference Management and Other Things
This change adds automatic Python refcounting using a Racket executor.  This
means when a Python pointer is about to be GC'd, racket will automatically call
Py_DecRef on it.

The FFI calls have been updated to call either "handle-new-ref" or
"handle-borrowed-ref" depending on if the Python API docs say the function
returns a new reference or a borrowed reference.  Additionally, functions which
steal a reference need to be wrapped to call Py_IncRef to ensure that
everything will still add up.

This change removes the call to Py_Finalize that would occur to tear down the
interpreter when reloading snake-oil in dr Racket etc, which is probably a bad
idea to do now.

In the other news, snake-oil now uses API calls that are part of the Stable
ABI, but are new as of Python 3.10, so that is now the minimum supported
version.

I have not exhaustively verified that this new scheme is correct, but it
appears to be functional from my very limited testing so far.